Classes for the Trip Advisor Scrap
It contains the classes for the Trip Advisor Scrap. This is not yet the most efficient scrap algorithm possible mainly due to the need for stronger libraries masking the IP address to avoid the Trip Advisor blockage. Yet, it is an excelent start point for obtaining thousands of info on a wide variety of searches.